rows per trigger and functions for press handlers
i'd be curious to determine the overhead of moving each branch into its
own press handler. i feel like it will make it easier to deal with
multiple menu modes and the individual display/press handling modules
could be easily extended as long as there's a description of how to
write to the global monome led display.